pyproject.toml
[project]
name = "agentic-lifelog"
version = "0.1.0"
description = "Agentic Lifelog - Personal Intelligence Platform for NVIDIA GTC 25 Hackathon"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.2.0",
    "langchain-nvidia-ai-endpoints>=0.3.0",
    "langgraph>=0.2.0",
    "streamlit>=1.40.0",
    "chromadb>=0.5.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
    "openai>=1.50.0",
    "typing-extensions>=4.12.0",
    "plotly>=5.18.0",
    "requests>=2.31.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

